+++ src/geodata/__init__.py
@@ -19,6 +19,7 @@ from .dataset import Dataset
 from .mask import Mask
 from .plot import *  # noqa: F403
 from .model import *  # noqa: F403
+from .pvlib import *  # noqa: F403
 
 __author__ = "Michael Davidson (UCSD), William Honaker"
 __copyright__ = "GNU GPL 3 license"
